+++ bootique-di/src/main/java/io/bootique/di/spi/OptionalBindingBuilder.java
@@ -1,5 +1,6 @@
 package io.bootique.di.spi;
 
+import java.util.Optional;
 import javax.inject.Provider;
 
 import io.bootique.di.Key;
@@ -14,7 +15,16 @@ class OptionalBindingBuilder<T> extends DefaultBindingBuilder<T> {
 
     @Override
     protected void initBinding() {
-        injector.putOptionalBinding(bindingKey, nullProvider());
+        Binding<T> binding = injector.getBinding(bindingKey);
+        // do not override existing binding with optional one
+        if(binding == null) {
+            injector.putOptionalBinding(bindingKey, nullProvider());
+        }
+        // add binding to Optional<T> type
+        injector.putBinding(Key.getOptionalOf(bindingKey), () -> {
+            T value = injector.getProvider(bindingKey).get();
+            return value == null ? Optional.empty() : Optional.of(value);
+        });
     }
 
     @SuppressWarnings("unchecked")